Clean code.

+++ DataMapper/SqlServerDAO/SQLBidDataServices.cs
@@ -2,14 +2,14 @@
 // Debu Matei
 // </copyright>
 
-using DataMapper.Interfaces;
-using DomainModel.Models;
-using log4net;
-using System.Data.Entity;
-using System.Diagnostics.CodeAnalysis;
-
 namespace DataMapper.SqlServerDAO
 {
+    using System.Data.Entity;
+    using System.Diagnostics.CodeAnalysis;
+    using DataMapper.Interfaces;
+    using DomainModel.Models;
+    using log4net;
+
     /// <summary>
     /// The bid data setvices.
     /// </summary>
@@ -18,71 +18,96 @@ namespace DataMapper.SqlServerDAO
     public class SQLBidDataServices : IBidDataServices
     {
         /// <summary>
-        /// The logger
+        /// The logger.
         /// </summary>
         private static readonly ILog Logger = LogManager.GetLogger(Environment.MachineName);
+
         /// <summary>
         /// Adds the bid.
         /// </summary>
         /// <param name="bid">The bid.</param>
-        /// <returns></returns>
+        /// <returns>bool.</returns>
         public bool AddBid(Bid bid)
         {
-           using(AuctionContext context = new AuctionContext())
+           using (AuctionContext context = new AuctionContext())
             {
                 try
                 {
-                    context.Products.Attach(bid.Product);
-                    context.Users.Attach(bid.Buyer);
-                    context.Bids.Add(bid);
-                    context.SaveChanges();
-
-                }catch(Exception exception)
+                    if (context.Products != null && context.Users != null && context.Bids != null)
+                    {
+                        context.Products.Attach(bid.Product);
+                        context.Users.Attach(bid.Buyer);
+                        context.Bids.Add(bid);
+                        context.SaveChanges();
+                    }
+                    else
+                    {
+                        throw new InvalidOperationException("One or more required DbSet properties are null.");
+                    }
+                }
+                catch (Exception exception)
                 {
-                    Logger.Error("Error while adding new bid: " + exception.Message.ToString() + " " + exception.InnerException.ToString());
+                    Logger.Error("Error while adding new bid: " + exception.Message.ToString());
                     return false;
                 }
             }
-            Logger.Info("Bid added successfully");
-            return true;
+
+           Logger.Info("Bid added successfully");
+           return true;
         }
 
         /// <summary>
         /// Deletes the bid.
         /// </summary>
         /// <param name="bid">The bid.</param>
-        /// <returns></returns>
+        /// <returns>bool.</returns>
         public bool DeleteBid(Bid bid)
         {
             using (AuctionContext context = new AuctionContext())
             {
                 try
                 {
-                    context.Bids.Attach(bid);
-                    context.Bids.Remove(bid);
-                    context.SaveChanges();
-                }catch(Exception exception)
+                    if (context.Bids != null)
+                    {
+                        context.Bids.Attach(bid);
+                        context.Bids.Remove(bid);
+                        context.SaveChanges();
+                    }
+                    else
+                    {
+                        throw new InvalidOperationException("The Bids are null");
+                    }
+                }
+                catch (Exception exception)
                 {
-                    Logger.Error("Error while deleting bid: " + exception.Message.ToString() + " " + exception.InnerException.ToString());
+                    Logger.Error("Error while deleting bid: " + exception.Message.ToString());
                     return false;
                 }
             }
+
             Logger.Info("Bid deleted successfully!");
             return true;
-
         }
 
         /// <summary>
         /// Gets all bids.
         /// </summary>
-        /// <returns></returns>
+        /// <returns>a list of bids.</returns>
         public IList<Bid> GetAllBids()
         {
             IList<Bid> bids = new List<Bid>();
-            using (AuctionContext context =  new AuctionContext())
+            using (AuctionContext context = new AuctionContext())
             {
-                bids = context.Bids.Include("Product").Include("Buyer").OrderBy((bid) => bid.Id).ToList();
+                if (context.Bids != null)
+                {
+                    bids = context.Bids.Include("Product").Include("Buyer").OrderBy((bid) => bid.Id).ToList();
+                }
+                else
+                {
+                    throw new InvalidOperationException("The Bids are null");
+                }
             }
+
             return bids;
         }
 
@@ -90,31 +115,45 @@ namespace DataMapper.SqlServerDAO
         /// Gets the bid by identifier.
         /// </summary>
         /// <param name="id">The identifier.</param>
-        /// <returns></returns>
+        /// <returns>a bid.</returns>
         public Bid GetBidById(int id)
         {
-            Bid bid = null;
+            Bid? bid = null;
 
-            using(AuctionContext context = new AuctionContext())
+            using (AuctionContext context = new AuctionContext())
             {
-               bid =  context.Bids.Include("Product").Include("Buyer").Where((bid) => bid.Id == id).FirstOrDefault();
+                if (context.Bids != null)
+                {
+                    bid = context.Bids.Include("Product").Include("Buyer").Where((bid) => bid.Id == id).FirstOrDefault();
+                }
+                else
+                {
+                    throw new InvalidOperationException("The Bids are null");
+                }
             }
 
-            return bid;
+            return bid!;
         }
 
         /// <summary>
         /// Gets the bids by product identifier.
         /// </summary>
         /// <param name="id">The identifier.</param>
-        /// <returns></returns>
+        /// <returns>a list of bids.</returns>
         public IList<Bid> GetBidsByProductId(int id)
         {
             IList<Bid> productBids = new List<Bid>();
 
             using (AuctionContext context = new AuctionContext())
             {
-                productBids = context.Bids.Include("Product").Include("Buyer").Where((bid) => bid.Product.Id == id).OrderByDescending((bid) => bid.DateAndTime).ToList();
+                if (context.Bids != null)
+                {
+                    productBids = context.Bids.Include("Product").Include("Buyer").Where((bid) => bid.Product.Id == id).OrderByDescending((bid) => bid.DateAndTime).ToList();
+                }
+                else
+                {
+                    throw new InvalidOperationException("The Bids are null");
+                }
             }
 
             return productBids;
@@ -124,15 +163,23 @@ namespace DataMapper.SqlServerDAO
         /// Gets the no of active bids by user identifier.
         /// </summary>
         /// <param name="id">The identifier.</param>
-        /// <returns></returns>
+        /// <returns>int.</returns>
         public int GetNoOfActiveBidsByUserId(int id)
         {
-            int activeBids;
+            int activeBids = 0;
 
             using (AuctionContext context = new AuctionContext())
             {
-                activeBids = context.Bids.Include("Product").Include("Buyer").Where((bid) => bid.Buyer.Id == id && bid.Product.TerminationDate > DateTime.Now).Select((bid) => bid.Product.Id).Distinct().Count();
+                if (context.Bids != null)
+                {
+                    activeBids = context.Bids.Include("Product").Include("Buyer").Where((bid) => bid.Buyer.Id == id && bid.Product.TerminationDate > DateTime.Now).Select((bid) => bid.Product.Id).Distinct().Count();
+                }
+                else
+                {
+                    throw new InvalidOperationException("The Bids are null");
+                }
             }
+
             return activeBids;
         }
 
@@ -140,22 +187,31 @@ namespace DataMapper.SqlServerDAO
         /// Updates the bid.
         /// </summary>
         /// <param name="bid">The bid.</param>
-        /// <returns></returns>
+        /// <returns>bool.</returns>
         public bool UpdateBid(Bid bid)
         {
-            using(AuctionContext context = new AuctionContext())
+            using (AuctionContext context = new AuctionContext())
             {
                 try
                 {
-                    context.Bids.Attach(bid);
-                    context.Entry(bid).State = EntityState.Modified;
-                    context.SaveChanges();
-                }catch(Exception exception)
+                    if (context.Bids != null)
+                    {
+                        context.Bids.Attach(bid);
+                        context.Entry(bid).State = EntityState.Modified;
+                        context.SaveChanges();
+                    }
+                    else
+                    {
+                        throw new InvalidOperationException("The Bids are null");
+                    }
+                }
+                catch (Exception exception)
                 {
-                    Logger.Error("Error while updating bid: " + exception.Message.ToString() + " " + exception.InnerException.ToString());
+                    Logger.Error("Error while updating bid: " + exception.Message.ToString());
                     return false;
                 }
             }
+
             Logger.Info("Bid update successfully!");
             return true;
         }
